ambulance_ml_model/requirements.txt
ultralytics
opencv-python
torch
torchvision
numpy
lap>=0.5.12
pyserial
streamlit